Support fix in facter 2.2 to lsbmajdistrelease
Facter 2.2 fixed the meaning of lsbmajdistrelease under Ubuntu
(so that e.g. Ubuntu 14.04.1 has lsbmajdistrelease = 14.04 rather
than 14).

To accomodate this fix, this commit changes the version check
to use a regex. The regex is also pessimistic wrt to versions
beyond 2014, so this would require a change when tested against
Ubuntu 15.04+.